pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tmdb"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
  "pillow",
  "requests",
]

[tool.pytest.ini_options]
addopts = [
  "-vv",
    "--strict-markers",
    "--import-mode=importlib",
]
pythonpath = [
  "."
]
